refactor(hook-logic): Refactor hook logic to use a centralized api
