+++ src/data/api.ts
@@ -1,0 +1,88 @@
+import { API_BASE_URL } from '../utils/constants';
+import { formatPublishedDate, getIsbnNumber } from '../utils/utils';
+import { bookApiSchema, detailedBookSchema } from './bookSchema.ts';
+import { Book, DetailedBook, SearchOption } from './types.ts';
+
+export const fetchBookById = async (id: string, signal?: AbortSignal): Promise<DetailedBook> => {
+  const response = await fetch(`${API_BASE_URL}/${id}`, { signal });
+  const data = await response.json();
+
+  const { id: bookId, volumeInfo, saleInfo } = await detailedBookSchema.validate(data);
+
+  const {
+    title,
+    imageLinks,
+    authors,
+    categories,
+    publishedDate,
+    industryIdentifiers,
+    publisher,
+    subtitle,
+    description,
+    pageCount,
+    averageRating,
+    language,
+    previewLink,
+  } = volumeInfo;
+
+  const isbn = getIsbnNumber(industryIdentifiers);
+
+  return {
+    id: bookId,
+    isbn: isbn || '',
+    title,
+    subtitle: subtitle || '',
+    description,
+    pageCount,
+    averageRating: averageRating || '',
+    language: language || '',
+    authors,
+    categories,
+    publishedDate: formatPublishedDate(publishedDate),
+    publisher,
+    previewLink,
+    industryIdentifiers,
+    thumbnail: imageLinks?.thumbnail,
+    saleability: saleInfo.saleability,
+    country: saleInfo.country,
+    buyLink: saleInfo.buyLink,
+    retailPrice: saleInfo.retailPrice,
+  };
+};
+
+export const searchBooks = async (searchTerm: string, signal?: AbortSignal): Promise<Book[]> => {
+  const response = await fetch(`${API_BASE_URL}?q=${encodeURIComponent(searchTerm)}`, {
+    signal,
+  });
+
+  const data = await response.json();
+  const { items } = await bookApiSchema.validate(data);
+
+  return items.map(({ id, volumeInfo }) => {
+    const { title, authors, publishedDate, categories, industryIdentifiers, imageLinks } =
+      volumeInfo;
+
+    const isbn = getIsbnNumber(industryIdentifiers);
+
+    return {
+      id,
+      isbn: isbn || '',
+      title,
+      authors,
+      categories,
+      publishedDate: formatPublishedDate(publishedDate),
+      thumbnail: imageLinks?.thumbnail || '',
+    };
+  });
+};
+
+export const fetchBookSuggestions = async (searchTerm: string): Promise<SearchOption[]> => {
+  const response = await fetch(`${API_BASE_URL}?q=${encodeURIComponent(searchTerm)}`);
+  const data = await response.json();
+  const { items } = await bookApiSchema.validate(data);
+
+  return items.map((item) => ({
+    id: item.id,
+    title: item.volumeInfo.title,
+  }));
+};
